+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.4.0] - 2018-10-11
+
+### Added
+- ConfigUtils.tsx to handle props and options.
+- all Tabulator options can be set as React Tabulator props.
+- example of editable table.
+
 ## [0.3.4] - 2018-10-10
 
 ### Added